Split apart stan/model/util.hpp
